+++ tests/avsc/namespace_duplicate_test/namespace_duplicate_test/common/__init__.py
@@ -1,4 +1,0 @@
-""" init file for file imports at namespace level """
-from namespace_duplicate_test.common.Common import Common
-
-from namespace_duplicate_test.common.TraceContext import TraceContext