+++ tests/avsc/namespace_duplicate_test/namespace_duplicate_test/common/enums/__init__.py
@@ -1,2 +1,0 @@
-""" init file for file imports at namespace level """
-from namespace_duplicate_test.common.enums.CommonEnum import CommonEnum