+++ tests/avsc/namespace_duplicate_test/namespace_duplicate_test/unique/__init__.py
@@ -1,4 +1,0 @@
-""" init file for file imports at namespace level """
-from namespace_duplicate_test.unique.Unique import Unique
-
-from namespace_duplicate_test.unique.UniqueEnum import UniqueEnum